Reconnecting no longer eats websocket data.
When reconnecting, the previous task is cancelled. However, it is
still consuming data from the threadsafe queue, which won't get
fully consumed until the queue has data on it. Then, a
curio.CancelledError is raised, causing the data to be lost
forever.

When the CancelledError is raised whilst retrieving from the
queue, it is caught instead. Once being caught, the data is
placed back onto the queue, before the task finally dies. This
means that the first packet (be it RESUME or IDENTIFY) doesn't
get consumed by the previous task, and is ready to be consumed
by the new task created after a reconnect.

Signed-off-by: SunDwarf <sun@veriny.tf>