Switch to using MappingProxy items instead of dict views.
This makes `get_` methods redundant - items should be looked up via
either `.get()` or direct item access on the relevant properties.

This allows easy O(1) lookup of items inside objects without the
possibility of the user overwriting data as would be the threat when
allowing direct access of the underlying dictionaries.

Signed-off-by: Laura <l@veriny.tf>